+++ platform/lxd.go
@@ -155,7 +155,7 @@ func initiateLxd(vm Lxd, whichLxc string) error {
 		whichLxc,
 		"network",
 		"create",
-		vm.Settings.Profile+"br0",
+		"bravebr0",
 		"ipv6.address=none",
 		bridge,
 		"ipv4.nat=true")
@@ -170,13 +170,13 @@ func initiateLxd(vm Lxd, whichLxc string) error {
 		whichLxc,
 		"network",
 		"attach-profile",
-		vm.Settings.Profile+"br0",
+		"bravebr0",
 		vm.Settings.Profile,
 		"eth0")
 	if err != nil {
 		_ = shared.ExecCommand(whichLxc, "profile", "delete", vm.Settings.Profile)
 		_ = shared.ExecCommand(whichLxc, "storage", "delete", vm.Settings.StoragePool.Name)
-		_ = shared.ExecCommand(whichLxc, "network", "delete", vm.Settings.Profile+"br0")
+		_ = shared.ExecCommand(whichLxc, "network", "delete", "bravebr0")
 
 		return errors.New("Failed to attach network to profile: " + err.Error())
 	}
