+++ shared/strings.go
@@ -14,7 +14,7 @@ Looks like you're initialising an already configured Bravetools environment.
 You can manually cleanup your Bravetools LXD configuration using LXD CLI:				
 1. lxc profile delete brave
 2. lxc storage delete brave-[timestamp]
-3. lxc network delete {$USER}br0
+3. lxc network delete bravebr0
 `
 
 // REMOVEMP ..
